converted variables back to byte casted variables to byte in constructor

+++ src/sk/itsovy/jackanin/goodSchool/Main.java
@@ -3,8 +3,7 @@ package sk.itsovy.jackanin.goodSchool;
 public class Main {
     public static void main(String[] args) {
         School school = new School();
-        Student student1 = new Student("Karol", "Novak", 1, 2, 1, 24);
+        Student student1 = new Student("Karol", "Novak", (byte)1, (byte)2, (byte)1, 24);
         school.addStudent(student1);
-
     }
 }
